refactoring day03
