+++ 2019/src/lib.rs
@@ -1,2 +1,4 @@
 pub mod day01;
 pub mod day02;
+pub mod day03;
+pub mod utils;
