+++ 2019/tests/day03_test.rs
@@ -1,0 +1,14 @@
+#[cfg(test)]
+mod tests {
+    use aoc2019::day03;
+
+    #[test]
+    fn part1_works() {
+        assert_eq!(280, day03::part1());
+    }
+
+    #[test]
+    fn part2_works() {
+        assert_eq!(10554, day03::part2());
+    }
+}
